+++ Model/WorkerData.php
@@ -1,16 +1,12 @@
 <?php
 
-
 namespace Leantime\Plugins\APIData\Model;
-
 
 readonly class WorkerData
 {
     public function __construct(
-        public int                  $id,
-        public string              $email,
-        public string              $name,
-    )
-    {
-    }
+        public int $id,
+        public ?string $email,
+        public ?string $name,
+    ) {}
 }